Lingering off_t problem
git-svn-id: http://oss.mars.asu.edu/svn/davinci/iomedley/trunk@949 dedd6328-c50f-0410-8be5-b54a49638459

+++ io_pnm.c
@@ -14,7 +14,7 @@ static int ReadPNMHeader(FILE *fp, int *xout, int *yout, int *zout, int *bits, i
 static int get_int(FILE *fp);
 static int getbit(FILE *fp);
 int iom_ReadPNM(FILE *fp, char *filename, int *xout, int *yout, 
-            int *zout, int *bits, void **dout, off_t *data_offset);
+            int *zout, int *bits, void **dout, int *data_offset);
 
 int
 iom_isPNM(FILE *fp)
